Some corrections using the new osutils.rename. **ALL TESTS PASS**

+++ bzrlib/tests/blackbox/test_too_much.py
@@ -211,17 +211,20 @@ class TestCommands(ExternalBase):
         os.rmdir('revertdir')
         self.runbzr('revert')
 
-        os.symlink('/unlikely/to/exist', 'symlink')
-        self.runbzr('add symlink')
-        self.runbzr('commit -m f')
-        os.unlink('symlink')
-        self.runbzr('revert')
-        self.failUnlessExists('symlink')
-        os.unlink('symlink')
-        os.symlink('a-different-path', 'symlink')
-        self.runbzr('revert')
-        self.assertEqual('/unlikely/to/exist',
-                         os.readlink('symlink'))
+        if has_symlinks():
+            os.symlink('/unlikely/to/exist', 'symlink')
+            self.runbzr('add symlink')
+            self.runbzr('commit -m f')
+            os.unlink('symlink')
+            self.runbzr('revert')
+            self.failUnlessExists('symlink')
+            os.unlink('symlink')
+            os.symlink('a-different-path', 'symlink')
+            self.runbzr('revert')
+            self.assertEqual('/unlikely/to/exist',
+                             os.readlink('symlink'))
+        else:
+            self.log("skipping revert symlink tests")
         
         file('hello', 'wt').write('xyz')
         self.runbzr('commit -m xyz hello')
